Update StartupManager.ps1
Added the -WindowStyle Hidden flag due to josh3003's suggestion of having the PowerShell window not show on startup. https://github.com/Correlander/BetterDiscordAutoInstaller/issues/1

Also added the -NoProfile flag as while searching for a flag/method to do it through PowerShell, I found that, and it seems best as that way if anyone with a customized environment wants to use this, they still can without conflict.